StremSence-videoM/app/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==5.0.1
blinker==1.9.0
boto3==1.35.78
botocore==1.35.78
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
Flask==3.1.0
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==3.0.2
minio==7.2.12
pycparser==2.22
pycryptodome==3.21.0
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
s3transfer==0.10.4
six==1.17.0
typing_extensions==4.12.2
urllib3==1.26.20
Werkzeug==3.1.3
zipp==3.21.0
